% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{create_general_plot_with_default_settings}
\alias{create_general_plot_with_default_settings}
\title{Create a generic ggplot object with default aesthetics}
\usage{
create_general_plot_with_default_settings()
}
\description{
Returns a ggplot with common aesthetics, like removed gridlines, grey axis lines etc.
}
