% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Column definitions of all datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
data_dictionary
}
\description{
This dataset provides metadata about all datasets in the package r2dii.ggplot.
}
\section{Definitions}{

\itemize{
\item \code{column} (character): The name of a dataset-column.
\item \code{dataset} (character): The name of a dataset.
\item \code{definition} (character): The definition of a dataset-column.
\item \code{typeof} (character): The result of \code{typeof()}, one of double, integer, logical, or character..
}
}

\examples{
head(data_dictionary)
}
\concept{meta}
\keyword{datasets}
