% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Imports example data set for plotting}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1170 rows and 8 columns.
}
\usage{
example_data
}
\value{
A dataframe.
}
\description{
The example data set has the same structure as the banks' data and can be
used to explore the plotting possibilities.
}
\examples{
example_data
}
\keyword{datasets}
