% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filter_data_for_techmix_chart}
\alias{filter_data_for_techmix_chart}
\title{Filters pre-processed data to be ready for plotting a techmix chart for a sector}
\usage{
filter_data_for_techmix_chart(
  data,
  sector,
  years,
  region,
  scenario_source,
  scenario,
  value_name
)
}
\arguments{
\item{data}{pre-processed input data}

\item{sector}{sector for which to filter the data (a character string)}

\item{years}{years which we want to plot in the graph (an array of integer values)}

\item{region}{region for which to filter the data (a character string)}

\item{scenario_source}{scenario source for which to filter the data (a character string)}

\item{scenario}{scenario to plot in the graph (a character string)}

\item{value_name}{the name of the value to be plotted as a bar chart (a character string)}
}
\description{
Filters pre-processed data to be ready for plotting a techmix chart for a sector
}
