% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filter_data_for_trajectory_chart}
\alias{filter_data_for_trajectory_chart}
\title{Filters pre-processed data to be ready for plotting a trajectory chart for a
technology}
\usage{
filter_data_for_trajectory_chart(
  data,
  sector,
  technology,
  region,
  scenario_source,
  value_name,
  end_year = 2025,
  normalize_to_start_year = TRUE
)
}
\arguments{
\item{data}{pre-processed input data}

\item{sector}{sector for which to filter the data (a character string)}

\item{technology}{technology for which to filter the data (a character
string)}

\item{region}{region for which to filter the data (a character string)}

\item{scenario_source}{scenario source for which to filter the data (a
character string)}

\item{value_name}{the name of the value to be plotted in the trajectory chart
(a character string)}

\item{end_year}{cut off year for the chart (an integer; default = 2025)}

\item{normalize_to_start_year}{flab indicating whether the values should be
normalized (boolean; default = TRUE)}
}
\description{
Filters pre-processed data to be ready for plotting a trajectory chart for a
technology
}
