% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_title.R
\name{format_label}
\alias{format_label}
\alias{recode_label}
\title{Format legend labels}
\usage{
format_label(x)

recode_label(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\code{format_label()} formats legend labels from the expected values of metrics in
\code{sda}-like and \code{market_share}-like datasets:
\itemize{
\item Scenarios are stripped of the suffix "target_" and converted to upper case.
\item Other metrics are converted to title case.
\item For all metrics any sequence of non alpha-numeric characters is replaced
with a single space.
\code{recode_label()} recoding metric values and outputs them in title case.
}
}
\examples{
format_label(c("corporate_economy", "target_sds"))
format_label(c("corporate_._economy", "target_sds_abc"))
recode_label(c("projected", "corporate_economy", "target_sds"))
}
\concept{helpers}
