% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Imports example data set for plotting}
\usage{
get_example_data()
}
\value{
A dataframe.
}
\description{
The example data set has the same structure as the banks' data and can be
used to explore the plotting possibilities.
}
\examples{
get_example_data()
}
