% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{get_r2dii_sector_colours}
\alias{get_r2dii_sector_colours}
\title{Get the predefined technology colors for a sector}
\usage{
get_r2dii_sector_colours(sector)
}
\arguments{
\item{sector}{sector for which we want to retrieve colors (a character string)}
}
\description{
Get the predefined technology colors for a sector
}
