% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{get_sector_colors}
\alias{get_sector_colors}
\title{Get the predefined technology colors for a sector}
\usage{
get_sector_colors(sector)
}
\arguments{
\item{sector}{sector for which we want to retrieve colors (a character string)}
}
\description{
Get the predefined technology colors for a sector
}
