% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\docType{data}
\name{market_share}
\alias{market_share}
\title{Dataset imitating the output of \code{r2dii.analysis::target_market_share()}}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1170 rows and 8 columns.
}
\usage{
market_share
}
\description{
Dataset imitating the output of \code{r2dii.analysis::target_market_share()}
}
\examples{
market_share
}
\seealso{
\code{\link[r2dii.analysis:target_market_share]{r2dii.analysis::target_market_share()}}.

Other example datasets: 
\code{\link{sda}}
}
\concept{example datasets}
\keyword{datasets}
