% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_match_success_rate.R
\name{plot_match_success_rate}
\alias{plot_match_success_rate}
\title{Title}
\usage{
plot_match_success_rate(.data, ..., .x = NULL)
}
\arguments{
\item{.data}{A data frame like the outputs of
\code{r2dii.match::calculate_match_success_rate()}.}

\item{...}{Additional grouping columns.}

\item{.x}{The name of the column to be used as the x-axis.}
}
\value{
An object of class "ggplot".
}
\description{
Title
}
\examples{
coverage <-
r2dii.data::loanbook_demo \%>\%
  r2dii.match::match_name(r2dii.data::abcd_demo) \%>\%
  r2dii.match::prioritize() \%>\%
  r2dii.match::calculate_match_success_rate(loanbook = r2dii.data::loanbook_demo)

plot_match_success_rate(coverage)

}
