% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_metareport_bubble}
\alias{plot_metareport_bubble}
\title{Create a meta-report bubble chart}
\usage{
plot_metareport_bubble(
  data,
  plot_title = NULL,
  x_title = "",
  y_title = "",
  investor_labels = NULL,
  colour_investors = TRUE,
  show_legend_when_coloured = TRUE
)
}
\arguments{
\item{data}{Dataframe with data processed for the chart. With columns:
investor_name, portfolio_name, value_x, value_y (dataframe).}

\item{plot_title}{Title of the plot; (character string).}

\item{x_title, y_title}{x- and y-axis title (character string).}

\item{investor_labels}{Dataframe with order and labels for investor types,
columns: investor_type, label. Can be used for preserving colour order
between different plots (using the same df will preserve the order). If no
is specified, investor_type from data is used as label. (dataframe).}

\item{colour_investors}{flag indicating if points should be coloured
(boolean).}

\item{show_legend_when_coloured}{flag indicating if legend should be showed
if points are coloured (boolean).}
}
\value{
An object of class "ggplot".
}
\description{
Create a meta-report bubble chart
}
\examples{
# TODO
}
