% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_metareport_map}
\alias{plot_metareport_map}
\title{Create a meta-report map chart}
\usage{
plot_metareport_map(
  data,
  plot_title = NULL,
  legend_title = NULL,
  sector = NULL
)
}
\arguments{
\item{data}{Dataframe with data processed for the chart. With columns: long,
lat, group, value (dataframe).}

\item{plot_title}{Title of the plot; (character string).}

\item{legend_title}{Title of the legend describing value units; (character
string).}

\item{sector}{Sector to be used for the map colouring. If none is specified,
default dark blue palette is used (character string)}
}
\value{
An object of class "ggplot".
}
\description{
Create a meta-report map chart
}
