% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_metareport_pacta_sectors}
\alias{plot_metareport_pacta_sectors}
\title{Create a small multiples bar chart with portfolio exposure to PACTA sectors}
\usage{
plot_metareport_pacta_sectors(
  data,
  bars_labels_specs = NULL,
  plot_title = "Percentage of Equity and Bonds Portfolios invested in PACTA sectors"
)
}
\arguments{
\item{data}{dataframe with data processed for the chart. With columns:
investor_name, asset_type, share_climate_relevant (dataframe)}

\item{bars_labels_specs}{(optional) dataframe with labels for investor
types, columns: investor_type, label. If no is specified, investor_type
from data is used as label. (dataframe; default = NULL)}

\item{plot_title}{title of the plot (character string; default = "Percentage
of Equity and Bonds Portfolios invested in PACTA sectors").}
}
\value{
an object of class "ggplot"
}
\description{
Create a small multiples bar chart with portfolio exposure to PACTA sectors
}
\examples{
# TODO
}
