% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_metareport_security_types}
\alias{plot_metareport_security_types}
\title{Create a bar chart with overview of asset types per investor type}
\usage{
plot_metareport_security_types(
  data,
  bars_asset_type_specs = data.frame(asset_type = c("Equity", "Bonds", "Others"), label
    = c("Equity", "Bonds", "Others"), r2dii_colour_name = c("dark_blue", "green",
    "grey")),
  bars_labels_specs = NULL
)
}
\arguments{
\item{data}{Dataframe with data processed for the chart. With columns:
investor_name, asset_type, share (dataframe).}

\item{bars_asset_type_specs}{Dataframe with specifications for each asset
type, columns: asset_type, legend label, r2dii_colour_name (dataframe).}

\item{bars_labels_specs}{Dataframe with labels for investor types, columns:
investor_type, label. If no is specified, investor_type from data is used
as label. (dataframe).}
}
\description{
This function plots a horizontal stacked bar chart with composition of
investors securities per investor type. No need to specify the colours if the
security types in your dataset are only: Equity, Bonds and Others. Otherwise
full specification needs to be passed to the function, for all security
types.
}
