% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_techmix_chart}
\alias{plot_techmix_chart}
\title{Create a techmix chart in a ggplot object}
\usage{
plot_techmix_chart(
  data,
  plot_title = "",
  show_legend = TRUE,
  df_tech_colours,
  df_bar_specs
)
}
\arguments{
\item{data}{filtered input data (dataframe with columns: technology,
metric_type, metric and value).}

\item{plot_title}{title of the plot (character string; default = "").}

\item{show_legend}{flag indicating whether legend should be shown (boolean;
default = TRUE).}

\item{df_tech_colours}{dataframe containing colors per technology (dataframe
with columns: technology, label, color).}

\item{df_bar_specs}{dataframe containing order of bars and their labels
(dataframe with columns: metric_type, label).}
}
\description{
The function returns a ggplot object containing a stacked bar chart showing a
technology mix for different categories (portfolio, scenario, benchmark,
etc.).
}
\examples{
# TODO create an example or copy-paste an existing one from README or a test.
}
