% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timelineX.R
\name{plot_timelineX}
\alias{plot_timelineX}
\title{Create a timeline plot}
\usage{
plot_timelineX(data, extrapolate = FALSE)
}
\arguments{
\item{data}{A data frame. Requirements:
\itemize{
\item The structure must be like \link{sda}.
\item The column \code{sector} must have a single value (e.g. "cement").
}}

\item{extrapolate}{Logical of length 1. \code{TRUE} extrapolates to match the
furthest value in the data set.}
}
\value{
An object of class "ggplot".
}
\description{
Create a timeline plot
}
\examples{
library(dplyr, warn.conflicts = FALSE)

# `data` must meet documented "Requirements"
data <- filter(sda, sector == "cement")
plot_timelineX(data)
}
\seealso{
\link{sda}.

Other functions with minimal arguments: 
\code{\link{plot_techmixX}()},
\code{\link{plot_trajectoryX}()}
}
\concept{functions with minimal arguments}
