% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_trajectory_chart}
\alias{plot_trajectory_chart}
\title{Create a trajectory alignment chart in a ggplot object}
\usage{
plot_trajectory_chart(
  data,
  plotTitle = "",
  xTitle = "",
  yTitle = "",
  annotateData = FALSE,
  scenario_specs_good_to_bad,
  mainLineMetric,
  additionalLineMetrics = data.frame()
)
}
\arguments{
\item{data}{filtered input data (dataframe with columns: year, metric_type, metric and value)}

\item{plotTitle}{title of the plot (character string; default = "")}

\item{xTitle}{title of the x-axis (character string; default = "")}

\item{yTitle}{title of the y-axis (character string; default = "")}

\item{annotateData}{flag indicating whether the data should be annotated (boolean; default = FALSE)}

\item{scenario_specs_good_to_bad}{dataframe containing scenario specifications like color or label, ordered from the most to least sustainable (dataframe with columns: scenario, label, color)}

\item{mainLineMetric}{dataframe containing information about metric that should be plotted as the main line (datframe with columns: metric, label)}

\item{additionalLineMetrics}{dataframe containing information about additional metrics that should be plotted as lines (datframe with columns: metric, label; default = data.frame())}
}
\description{
The function returns a ggplot object containing a stacked bar chart showing a technology mix for different categories (portfolio, scenario, benchmark, etc.)
}
