% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_techmix.R
\name{prep_techmix}
\alias{prep_techmix}
\alias{prep_techmixB}
\title{Prepares pre-processed data for plotting a tech-mix chart}
\usage{
prep_techmix(
  data,
  sector_filter = c("automotive", "aviation", "cement", "oil and gas", "shipping",
    "steel", "power"),
  years_filter = NULL,
  region_filter = "global",
  scenario_source_filter = NULL,
  scenario_filter = NULL,
  value_to_plot = "technology_share"
)

prep_techmixB(data, value = "technology_share")
}
\arguments{
\item{data}{Pre-processed input data.}

\item{sector_filter}{Sector for which to filter the data (character string).}

\item{years_filter}{Years to plot in the graph (array of integer values).}

\item{region_filter}{Region for which to filter the data (character string).}

\item{scenario_source_filter}{Scenario source for which to filter the data
(character string).}

\item{scenario_filter}{Scenario to plot in the graph (character string).}

\item{value_to_plot, value}{The name of the value to be plotted as a bar chart
(character string).}
}
\description{
Prepares pre-processed data for plotting a tech-mix chart
}
\examples{
prep_techmix(
  market_share,
  sector_filter = "power",
  years_filter = c(2020, 2025),
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  scenario_filter = "sds",
  value_to_plot = "technology_share"
)
library(dplyr)

data <- market_share \%>\%
  filter(
    dplyr::between(year, 2020, 2025),
    scenario_source == "demo_2020",
    sector == "power",
    region == "global",
    metric \%in\% c("projected", "corporate_economy", "target_sds")
  )

prep <- prep_techmixB(data)
}
