% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_timeline.R
\name{prep_timeline}
\alias{prep_timeline}
\title{Prepare the output of \code{r2dii.analysis::target_sda()} for \code{plot_timeline()}}
\usage{
prep_timeline(data, extrapolate = FALSE)
}
\arguments{
\item{data}{Dataframe like the output of \code{r2dii.analysis::target_sda()}.}

\item{extrapolate}{Logical of length 1. \code{TRUE} extrapolates to match the
furthest value in the data set.}
}
\value{
Dataframe with columns: year, line_name, value, extrapolated.
}
\description{
Prepare the output of \code{r2dii.analysis::target_sda()} for \code{plot_timeline()}
}
\examples{
tail(prep_timeline(sda))
tail(prep_timeline(sda, extrapolate = TRUE))
}
\seealso{
\link{sda}.
}
