% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_trajectoryY.R
\name{prep_trajectoryY}
\alias{prep_trajectoryY}
\title{Prepare the output of \code{r2dii.analysis::target_market_share()} for \code{plot_trajectory()}}
\usage{
prep_trajectoryY(
  data,
  sector_filter,
  technology_filter,
  region_filter,
  scenario_source_filter,
  value = "production",
  metric = "metric",
  end_year_filter = 2025,
  normalize = TRUE
)
}
\arguments{
\item{technology_filter}{String of length 1. Technology to pick from the
\code{data}.}

\item{end_year_filter}{Numeric of length 1. Cut-off year for the plot.}

\item{normalize}{Logical of length-1. \code{TRUE} normalizes to the start year.}
}
\value{
A data frame.
}
\description{
Prepare the output of \code{r2dii.analysis::target_market_share()} for \code{plot_trajectory()}
}
\examples{
prep_trajectoryY(
  market_share,
  sector_filter = "power",
  technology_filter = "oilcap",
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  value = "production"
)
}
