% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_metareport_distribution_chart}
\alias{prepare_for_metareport_distribution_chart}
\title{Prepared results data per asset type for distribution chart plot}
\usage{
prepare_for_metareport_distribution_chart(
  data_asset_type,
  sector_filter,
  technologies_filter,
  year_filter,
  value_to_plot = "plan_carsten"
)
}
\arguments{
\item{data_asset_type}{dataframe in the shape of
"Equity/Bonds_results_portfolio.rda" dataset from PACTA analysis output in
"30_Results" folder (dataframe)}

\item{sector_filter}{sector to be used for filtering (a character string)}

\item{technologies_filter}{technologies to be used for filtering (a vector
of charater strings)}

\item{year_filter}{year to be used for filtering (integer)}

\item{value_to_plot}{variable to be used as value for plotting (a character
string)}
}
\value{
dataframe with columns investor_name, portfolio_name, value
}
\description{
Prepared results data per asset type for distribution chart plot
}
\examples{
# TODO
}
