% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_metareport_pacta_sectors_mix_chart}
\alias{prepare_for_metareport_pacta_sectors_mix_chart}
\title{Aggregates and prepares PACTA analysis overview data to be an input for
meta-report PACTA sectors mix chart}
\usage{
prepare_for_metareport_pacta_sectors_mix_chart(data_overview)
}
\arguments{
\item{data_overview}{dataframe in the shape of
".._overview_portfolio.rda" data set from PACTA analysis output in
"30_Processed_Inputs" folder (dataframe)}
}
\value{
dataframe prepared for the plot
}
\description{
Aggregates and prepares PACTA analysis overview data to be an input for
meta-report PACTA sectors mix chart
}
\examples{
# TODO
}
