% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_metareport_security_type_chart}
\alias{prepare_for_metareport_security_type_chart}
\title{Prepares PACTA total_portfolio data for meta-report security type bar chart}
\usage{
prepare_for_metareport_security_type_chart(
  data_total_portfolio,
  other_asset_types = c("Funds", "Others", "Unclassifiable")
)
}
\arguments{
\item{data_total_portfolio}{Dataframe in the shape of
".._total_portfolio.rda" dataset from PACTA analysis output in
"30_Processed_Inputs" folder (dataframe).}

\item{other_asset_types}{Array of character strings that should be summed up
as "Other" asset type (array of character strings).}
}
\description{
This function aggregates and prepares one of PACTA analysis result files
".._total_portfolio.rda" from "30_Processed_Inputs" folder to form an input
that can be used for plotting meta-report security type coverage per investor
type bar chart.
}
