% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_techmix_chart}
\alias{prepare_for_techmix_chart}
\title{Prepares pre-processed data for plotting a tech-mix chart}
\usage{
prepare_for_techmix_chart(
  data_preprocessed,
  sector_filter,
  years_filter,
  region_filter,
  scenario_source_filter,
  scenario_filter,
  value_name
)
}
\arguments{
\item{data_preprocessed}{Pre-processed input data.}

\item{sector_filter}{Sector for which to filter the data (character string).}

\item{years_filter}{Years to plot in the graph (array of integer values).}

\item{region_filter}{Region for which to filter the data (character string).}

\item{scenario_source_filter}{Scenario source for which to filter the data
(character string).}

\item{scenario_filter}{Scenario to plot in the graph (character string).}

\item{value_name}{The name of the value to be plotted as a bar chart
(character string).}
}
\description{
Prepares pre-processed data for plotting a tech-mix chart
}
