% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_trajectory_chart}
\alias{prepare_for_trajectory_chart}
\title{Prepares pre-processed data to be ready for plotting a trajectory chart for a
technology}
\usage{
prepare_for_trajectory_chart(
  data_preprocessed,
  sector_filter,
  technology_filter,
  region_filter,
  scenario_source_filter,
  value_name,
  end_year_filter = 2025,
  normalize_to_start_year = TRUE
)
}
\arguments{
\item{data_preprocessed}{pre-processed input data}

\item{sector_filter}{sector for which to filter the data (a character string)}

\item{technology_filter}{technology for which to filter the data (a character
string)}

\item{region_filter}{region for which to filter the data (a character string)}

\item{scenario_source_filter}{scenario source for which to filter the data (a
character string)}

\item{value_name}{the name of the value to be plotted in the trajectory chart
(a character string)}

\item{end_year_filter}{cut off year for the chart (an integer; default = 2025)}

\item{normalize_to_start_year}{flab indicating whether the values should be
normalized (boolean; default = TRUE)}
}
\description{
Prepares pre-processed data to be ready for plotting a trajectory chart for a
technology
}
