% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_data.R
\name{process_input_data}
\alias{process_input_data}
\title{Performs the initial processing on raw input data in banks' format}
\usage{
process_input_data(data)
}
\arguments{
\item{data}{Raw input data in the format of banks' output.}
}
\value{
A dataframe with additional column \code{metric_type} and modified
\code{metric}.
}
\description{
This function processes the data in banks' format so that it can
be used later in data filtering functions for charts.
}
\examples{
data <- example_data
process_input_data(data)
}
