% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_techmix_explorer.R
\name{qplot_techmix_explorer}
\alias{qplot_techmix_explorer}
\title{Interactive \code{qplot_techmix} options explorer}
\usage{
qplot_techmix_explorer(data)
}
\arguments{
\item{data}{A data frame like the output of \code{prep_techmix()}.}
}
\value{
An object of class "ggplot".
}
\description{
An interactive shiny widget to explore the \code{qplot_techmix} options.
}
\examples{
qplot_techmix_explorer(market_share)

}
\seealso{
plot_techmix
}
