% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_trajectory_explorer.R
\name{qplot_trajectory_explorer}
\alias{qplot_trajectory_explorer}
\title{Interactive \code{qplot_trajectory} options explorer}
\usage{
qplot_trajectory_explorer(data)
}
\arguments{
\item{data}{A data frame like the outputs of \code{prep_trajectory()}.
\itemize{
\item (Optional) If present, the column \code{label} is used for data labels.
}}
}
\value{
An object of class "ggplot".
}
\description{
An interactive shiny widget to explore the \code{qplot_trajectory} options.
}
\examples{
qplot_trajectory_explorer(market_share)

}
\seealso{
plot_trajectory
}
