% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii_colours.R
\docType{data}
\name{r2dii_colours}
\alias{r2dii_colours}
\alias{palette_colours}
\alias{scenario_colours}
\alias{sector_colours}
\alias{technology_colours}
\title{Colour datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 3 columns.
}
\usage{
palette_colours

scenario_colours

sector_colours

technology_colours
}
\description{
All datasets have at least two columns:
\itemize{
\item \code{label}: Text label of the colour.
\item \code{hex}: Hex code of the colour.
}
}
\details{
In \code{scenario_colours}, colours are ordered from red to green to be used in
trajectory charts.
}
\examples{
palette_colours

scenario_colours

sector_colours

technology_colours
}
\seealso{
Other datasets: 
\code{\link{market_share_demo}},
\code{\link{sda_demo}}
}
\concept{datasets}
\keyword{datasets}
