% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii_scenario_colours.R
\docType{data}
\name{r2dii_scenario_colours}
\alias{r2dii_scenario_colours}
\title{Dataset giving the colour palette for scenarios}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\source{
\url{https://github.com/2DegreesInvesting/r2dii.plot.static/issues/151}.
}
\usage{
r2dii_scenario_colours
}
\description{
Scenario colours ordered from red to green to be used in trajectory charts.
}
\details{
Column definitions:
\itemize{
\item \code{label}: Text label of the colour.
\item \code{hex}: Hex code of the colour.
}
}
\examples{
r2dii_scenario_colours
}
\seealso{
Other colour datasets: 
\code{\link{get_r2dii_technology_colours}()},
\code{\link{r2dii_palette_colours}()},
\code{\link{r2dii_sector_colours}()}
}
\concept{colour datasets}
\keyword{datasets}
