% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2dii_colours.R
\name{r2dii_sector_colours}
\alias{r2dii_sector_colours}
\title{Get the 2DII sector colour palette}
\usage{
r2dii_sector_colours()
}
\value{
Tribble with sector labels and colour hex strings.
}
\description{
Get the 2DII sector colour palette
}
\examples{
r2dii_sector_colours()
}
\seealso{
Other colour datasets: 
\code{\link{get_r2dii_technology_colours}()},
\code{\link{r2dii_palette_colours}()},
\code{\link{r2dii_scenario_colours}}
}
\concept{colour datasets}
