% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_r2dii_sector.R
\name{scale_colour_r2dii_sector}
\alias{scale_colour_r2dii_sector}
\alias{scale_color_r2dii_sector}
\alias{scale_fill_r2dii_sector}
\title{Custom 2DII sector colour and fill scales}
\usage{
scale_colour_r2dii_sector(sectors = NULL, ...)

scale_fill_r2dii_sector(sectors = NULL, ...)
}
\arguments{
\item{sectors}{A character vector. Specifies sector colours to use and their
order. Run \code{unique(r2dii.plot:::sector_colours$label)} to see available
labels. Similar to \code{value} parameter in \code{ggplot::scale_colour_manual()}.}

\item{...}{Other parameters passed on to \code{ggplot2::discrete_scale()}.}
}
\value{
An object of class "ScaleDiscrete".
}
\description{
A custom discrete colour and fill scales with colours from 2DII sector
palette.
}
\examples{
library(ggplot2, warn.conflicts = FALSE)

ggplot(data = mpg) +
 geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
 scale_colour_r2dii_sector()

 ggplot(data = mpg) +
 geom_histogram(mapping = aes(x = cyl, fill = class), position = "dodge", bins = 5) +
 scale_fill_r2dii_sector()
}
