% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.R
\docType{data}
\name{sda}
\alias{sda}
\title{Dataset imitating the output of \code{r2dii.analysis::target_sda()}}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 208 rows and 4 columns.
}
\source{
\url{https://github.com/2DegreesInvesting/r2dii.plot.static/issues/55}.
}
\usage{
sda
}
\description{
Example data set of emission factors for different metric types: projected,
benchmark and scenarios and different sectors.
}
\examples{
sda
}
\seealso{
\code{\link[r2dii.analysis:target_sda]{r2dii.analysis::target_sda()}}.

Other example datasets: 
\code{\link{market_share}}
}
\concept{example datasets}
\keyword{datasets}
