% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_label.R
\name{to_title}
\alias{to_title}
\alias{spell_out_technology}
\alias{format_label_techmix}
\title{Functions to replicate labels produced with \verb{qplot_*()} functions}
\usage{
to_title(x)

spell_out_technology(x)

format_label_techmix(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\itemize{
\item \code{to_title()} converts a string to title case. It's similar to but not
identical to \code{\link[tools:toTitleCase]{tools::toTitleCase()}}. It replaces a sequence of non
alpha-numeric characters to a single space, and applies title case to the
remaining words.
\item \code{spell_out_technology()} replaces technology abbreviations with their
spelled-out version.
\item \code{format_label_techmix()} replaces "projected" with "portfolio",
"targets*" (e.g. "target_sds") with "Scenario", and everything else with
"Benchmark".
}
}
\examples{
to_title(c("a.string", "ANOTHER_string"))
to_title(c("a.string", "another_string", "b.STRING"))

spell_out_technology(c("FIXME_hdv", "FIXMEcap", "FIXMEice"))

format_label_techmix(c("projected", "target_xyz", "whatever"))
# Weird case
format_label_techmix(c("corporate_._economy", "target_sds_abc"))
}
