% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2dii_colours.R
\name{get_r2dii_technology_colours}
\alias{get_r2dii_technology_colours}
\title{Get the predefined technology colors for a sector}
\usage{
get_r2dii_technology_colours(sector)
}
\arguments{
\item{sector}{Sector for which we want to retrieve colors (character string).}
}
\description{
Get the predefined technology colors for a sector
}
\examples{
get_r2dii_technology_colours("power")
}
\seealso{
Other colour datasets: 
\code{\link{r2dii_palette_colours}()},
\code{\link{r2dii_scenario_colours}},
\code{\link{r2dii_sector_colours}()}
}
\concept{colour datasets}
