% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emission_intensity.R
\name{plot_emission_intensity}
\alias{plot_emission_intensity}
\title{Create an emission intensity plot}
\usage{
plot_emission_intensity(data)
}
\arguments{
\item{data}{A data frame like the output of \code{prep_emission_intensity()}.}
}
\value{
An object of class "ggplot".
}
\description{
Create an emission intensity plot
}
\examples{
# plot with `qplot_emission_intensity()` parameters
data <- subset(sda_demo, sector == "cement" & region == "global")
data <- prep_emission_intensity(data, span_5yr = TRUE, convert_label = to_title)

plot_emission_intensity(data)
}
\seealso{
\link{sda_demo}.
}
