% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emission_intensity.R
\name{plot_emission_intensity}
\alias{plot_emission_intensity}
\title{Create an emission intensity plot}
\usage{
plot_emission_intensity(data)
}
\arguments{
\item{data}{A data frame. Requirements:
\itemize{
\item The structure must be like \link{sda}.
\item The column \code{sector} must have a single value (e.g. "cement").
}}
}
\value{
An object of class "ggplot".
}
\description{
Create an emission intensity plot
}
\examples{
# `data` must meet documented "Requirements"
data <- subset(sda, sector == "cement")
plot_emission_intensity(data)
}
\seealso{
\link{sda}.
}
