% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_metareport_distribution}
\alias{plot_metareport_distribution}
\title{Create a meta-report distribution chart}
\usage{
plot_metareport_distribution(
  data,
  plot_title = "",
  x_title = "",
  y_title = "",
  investor_labels = NULL
)
}
\arguments{
\item{data}{Dataframe with data processed for the chart. With columns:
investor_name, portfolio_name, value (dataframe).}

\item{plot_title}{Title of the plot; (character string).}

\item{x_title, y_title}{x- and y-axis title (character string).}

\item{investor_labels}{Dataframe with order and labels for investor types,
columns: investor_type, label. Can be used for preserving colour order
between different plots (using the same df will preserve the order). If no
is specified, investor_type from data is used as label. (dataframe).}
}
\value{
An object of class "ggplot".
}
\description{
Create a meta-report distribution chart
}
\examples{
# TODO
}
