% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_metareport_pacta_sectors}
\alias{plot_metareport_pacta_sectors}
\title{Create a small multiples bar chart with portfolio exposure to PACTA sectors}
\usage{
plot_metareport_pacta_sectors(
  data,
  bars_labels_specs = NULL,
  plot_title = NULL
)
}
\arguments{
\item{data}{Dataframe with data processed for the chart. With columns:
investor_name, asset_type, share_climate_relevant (dataframe).}

\item{bars_labels_specs}{Dataframe with labels for investor types, columns:
investor_type, label. If no is specified, investor_type from data is used
as label. (dataframe).}

\item{plot_title}{Title of the plot; if no title is specified, the plot title
is set in the code to a predetermined character string (character string).}
}
\value{
An object of class "ggplot".
}
\description{
Create a small multiples bar chart with portfolio exposure to PACTA sectors
}
\examples{
# TODO
}
