% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_metareport_pacta_sectors_mix}
\alias{plot_metareport_pacta_sectors_mix}
\title{Create a small multiples stacked bar chart with percentage investment in
PACTA sectors}
\usage{
plot_metareport_pacta_sectors_mix(
  data,
  plot_title = NULL,
  df_sectors_order = data.frame(sector = c("steel", "cement", "shipping", "aviation",
    "automotive", "power", "coal", "oil&gas"), label = c("Steel", "Cement", "Shipping",
    "Aviation", "Automotive", "Power", "Coal", "Oil & Gas")),
  bars_labels_specs = NULL
)
}
\arguments{
\item{data}{Dataframe with data processed for the chart. With columns:
investor_name, asset_type, sector, share (dataframe).}

\item{plot_title}{Title of the plot; if no title is specified, the plot title
is set in the code to a predetermined character string (character string).}

\item{df_sectors_order}{Dataframe with ordered sector names and their labels
(dataframe).}

\item{bars_labels_specs}{Dataframe with labels for investor types, columns:
investor_type, label. If no is specified, investor_type from data is used
as label. (dataframe).}
}
\value{
An object of class "ggplot".
}
\description{
Create a small multiples stacked bar chart with percentage investment in
PACTA sectors
}
\examples{
# TODO
}
