% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_techmix.R
\name{plot_techmix}
\alias{plot_techmix}
\title{Create a techmix chart in a ggplot object}
\usage{
plot_techmix(data, df_bar_specs, tech_colours = NULL)
}
\arguments{
\item{data}{Filtered input data (dataframe with columns: technology,
metric_type, metric and value).}

\item{df_bar_specs}{Dataframe containing order of bars and their labels
(dataframe with columns: metric_type, label).}

\item{tech_colours}{Dataframe containing colours per technology (dataframe
with columns: technology, label(optional), colour).}
}
\description{
The function returns a ggplot object containing a stacked bar chart showing a
technology mix for different categories (portfolio, scenario, benchmark,
etc.).
}
\examples{
data <- prepare_for_techmix_chart(
  process_input_data(get_example_data()),
  sector_filter = "power",
  years_filter = c(2020, 2025),
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  scenario_filter = "sds",
  value_to_plot = "technology_share"
)
bar_specs <- dplyr::tibble(
  metric_type = c(
    "portfolio_2020",
    "benchmark_2020",
    "portfolio_2025",
    "benchmark_2025",
    "scenario_2025"
  ),
  label = c(
    "Portfolio 2020",
    "Benchmark 2020",
    "Portfolio 2025",
    "Benchmark 2025",
    "Target SDS 2025"
  )
)

print(
  plot_techmix(
    data,
    df_bar_specs = bar_specs
  )
)
}
