% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_techmix.R
\name{plot_techmix}
\alias{plot_techmix}
\title{Create a techmix chart in a ggplot object}
\usage{
plot_techmix(
  data,
  metric_type_order = NULL,
  metric_type_labels = NULL,
  tech_colours = NULL
)
}
\arguments{
\item{data}{Filtered input data, with columns: technology, metric_type,
metric and value.}

\item{metric_type_order}{Vector with the order of bars based on 'metric_type'
values.}

\item{metric_type_labels}{Vector with labels for bars. Order must follow that
in 'metric type order'.}

\item{tech_colours}{Dataframe containing colours per technology, with
columns: technology, colour, label(optional).}
}
\description{
The function returns a ggplot object containing a stacked bar chart showing a
technology mix for different categories (portfolio, scenario, benchmark,
etc.).
}
\examples{
data <- prep__techmix_chart(
  process_input_data(example_data),
  sector_filter = "power",
  years_filter = c(2020, 2025),
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  scenario_filter = "sds",
  value_to_plot = "technology_share"
)
print(
  plot_techmix(data)
)
}
