% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Create a time line plot}
\usage{
plot_timeline(
  data,
  lines_specs = NULL,
  plot_title = NULL,
  x_title = "Year",
  y_title = "Value"
)
}
\arguments{
\item{data}{Pre-processed data for the chart, with columns: year, value,
line_name.}

\item{lines_specs}{Dataframe containing order of lines, their labels and
(optionally) colour names from the r2dii_colours palette.}

\item{plot_title}{Title of the plot.}

\item{x_title, y_title}{x- and y-axis title.}
}
\value{
An object of class "ggplot".
}
\description{
Create a time line plot
}
\examples{
data_sda_cement <- prepare_for_timeline(sda_target,
  sector_filter = "cement",
  year_start = 2020,
  year_end = 2026,
  column_line_names = "emission_factor_metric",
  value_to_plot = "emission_factor_value"
)

lines_specs <- data.frame(
  "line_name" = c("projected", "corporate_economy", "target_demo", "adjusted_scenario_demo"),
  "label" = c("Projected", "Corporate Economy", "Target Demo", "Adjusted Scenario Demo"),
  "r2dii_colour_name" = c("dark_blue", "green", "grey", "orange")
)

plot_timeline(data_sda_cement,
  lines_specs = lines_specs,
  plot_title = "Emission intensity trend for Cement.",
  x_title = "Year",
  y_title = "Tons of CO2 per ton"
)
}
