% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Create a trajectory alignment chart in a ggplot object}
\usage{
plot_trajectory(
  data,
  plot_title = "",
  x_title = "",
  y_title = "",
  annotate_data = FALSE,
  scenario_specs_good_to_bad,
  main_line_metric,
  additional_line_metrics = data.frame()
)
}
\arguments{
\item{data}{Filtered input data (dataframe with columns: year, metric_type,
metric and value).}

\item{plot_title}{Title of the plot (character string).}

\item{x_title, y_title}{Title of the x- and y-axis (character string).}

\item{annotate_data}{Flag indicating whether the data should be annotated
(boolean).}

\item{scenario_specs_good_to_bad}{Dataframe containing scenario
specifications like color or label, ordered from the most to least
sustainable (dataframe with columns: scenario, label, color).}

\item{main_line_metric}{Dataframe containing information about metric that
should be plotted as the main line (dataframe with columns: metric, label).}

\item{additional_line_metrics}{Dataframe containing information about
additional metrics that should be plotted as lines (dataframe with columns:
metric, label).}
}
\description{
The function returns a ggplot object containing a stacked bar chart showing a
technology mix for different categories (portfolio, scenario, benchmark,
etc.).
}
