% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Create a trajectory plot}
\usage{
plot_trajectory(data, perc_y_scale = FALSE)
}
\arguments{
\item{data}{A data frame. Requirements:
\itemize{
\item Must have columns: \code{value}, \code{metric}, \code{year}, \code{sector}, \code{technology},
\code{region}, \code{scenario_source}.
\item depending on technology must have a \code{value_low} (brown) and \code{value_high}
column (see \code{r2dii.data::green_or_brown} for a classification)
\item The following columns must be present and have a single value: \code{sector},
\code{technology}, \code{region}, \code{scenario_source}.
\item (Optional) If present, the column \code{label} is used for data labels.
}}

\item{perc_y_scale}{Logical. \code{FALSE} defaults to using no label conversion.
Use \code{TRUE} to convert labels on y-axis to percentage using
\code{scales::percent} (the default behavior of \code{qplot_trajectory()}).}
}
\value{
An object of class "ggplot".
}
\description{
Create a trajectory plot
}
\examples{
# `data` must meet "Requirements" documented in `prep_trajectory`
data <- subset(
  market_share,
  sector == "power" &
    technology == "renewablescap" &
    region == "global" &
    scenario_source == "demo_2020"
)

prep_trajectory(data) \%>\%
plot_trajectory()
}
\seealso{
\link{prep_trajectory}.
}
