% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Create a trajectory plot}
\usage{
plot_trajectory(data, normalize = TRUE, main_line = NULL)
}
\arguments{
\item{data}{A data frame. Requirements:
\itemize{
\item The structure must be like \link{market_share}.
\item The following columns must have a single value: \code{sector}, \code{technology},
\code{region}, \code{scenario_source}.
}}

\item{normalize}{Logical of length-1. \code{TRUE} normalizes to the start year.}

\item{main_line}{String of length 1. The \code{metric} to plot as the line with
the most visual salience (solid black line). \code{NULL} defaults to
"projected".}
}
\value{
An object of class "ggplot".
}
\description{
Create a trajectory plot
}
\examples{
# `data` must meet documented "Requirements"
data <- subset(
  market_share,
  sector == "power",
  technology == "renewablescap",
  region == "global",
  scenario_source == "demo_2020"
)

plot_trajectory(data)

plot_trajectory(data, normalize = FALSE)
}
\seealso{
\link{market_share}.
}
