% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2diicharts.R
\name{plot_trajectory_chart}
\alias{plot_trajectory_chart}
\title{Create a trajectory alignment chart in a ggplot object}
\usage{
plot_trajectory_chart(
  data,
  plot_title = "",
  x_title = "",
  y_title = "",
  annotate_data = FALSE,
  scenario_specs_good_to_bad,
  main_line_metric,
  additional_line_metrics = data.frame()
)
}
\arguments{
\item{data}{filtered input data (dataframe with columns: year, metric_type, metric and value)}

\item{plot_title}{title of the plot (character string; default = "")}

\item{x_title}{title of the x-axis (character string; default = "")}

\item{y_title}{title of the y-axis (character string; default = "")}

\item{annotate_data}{flag indicating whether the data should be annotated (boolean; default = FALSE)}

\item{scenario_specs_good_to_bad}{dataframe containing scenario specifications like color or label, ordered from the most to least sustainable (dataframe with columns: scenario, label, color)}

\item{main_line_metric}{dataframe containing information about metric that should be plotted as the main line (dataframe with columns: metric, label)}

\item{additional_line_metrics}{dataframe containing information about additional metrics that should be plotted as lines (dataframe with columns: metric, label; default = data.frame())}
}
\description{
The function returns a ggplot object containing a stacked bar chart showing a technology mix for different categories (portfolio, scenario, benchmark, etc.)
}
