% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectoryA}
\alias{plot_trajectoryA}
\title{Create a trajectory alignment chart in a ggplot object}
\usage{
plot_trajectoryA(
  data,
  scenario_specs_good_to_bad,
  main_line_metric,
  additional_line_metrics = NULL
)
}
\arguments{
\item{data}{Filtered input data; with columns: year, metric_type, metric and
value.}

\item{scenario_specs_good_to_bad}{Data frame containing scenario
specifications like name label, ordered from the most to least sustainable;
with columns: scenario, label.}

\item{main_line_metric}{Data frame containing information about metric that
should be plotted as the main line; with columns: metric, label.}

\item{additional_line_metrics}{Data frame containing information about
additional metrics that should be plotted as lines; with columns: metric,
label).}
}
\description{
We are exploring different interfaces before release. We are
keen to hear feedback from beta-testers like you. Please try these
alternative interfaces and let us know which one you prefer. The main
difference between them is the number of arguments and how the input data
is used:
\itemize{
\item \code{plot_trajectoryA()} requires input arguments such as (at the minimum)
\code{scenario_specs_good_to_bad} and \code{main_line_metric} for specifying the
order and labels of scenario data and trajectory lines.
}
}
\examples{
# `plot_trajectoryA()` -------------------------------------------------------
data <- prep_trajectory(
  market_share,
  sector_filter = "power",
  technology_filter = "renewablescap",
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  value = "production"
)

scenario_specs <- dplyr::tibble(
  scenario = c("sds", "sps", "cps"),
  label = c("SDS", "STEPS", "CPS")
)

main_line_metric <- dplyr::tibble(metric = "projected", label = "Portfolio")

additional_line_metrics <- dplyr::tibble(
  metric = "corporate_economy",
  label = "Corporate Economy"
)

p <- plot_trajectoryA(data,
  scenario_specs_good_to_bad = scenario_specs,
  main_line_metric = main_line_metric,
  additional_line_metrics = additional_line_metrics
)

p
}
\seealso{
Other plotting functions: 
\code{\link{plot_techmix}()},
\code{\link{theme_2dii}()}
}
\concept{plotting functions}
