% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectoryX.R
\name{plot_trajectoryX}
\alias{plot_trajectoryX}
\title{Create a trajectory plot}
\usage{
plot_trajectoryX(data, normalize = TRUE, main_line = NULL)
}
\arguments{
\item{data}{A data frame. Requirements:
\itemize{
\item The structure must be like \link{market_share}.
\item The following columns must have a single value: \code{sector}, \code{technology},
\code{region}, \code{scenario_source}.
}}

\item{normalize}{Logical of length-1. \code{TRUE} normalizes to the start year.}

\item{main_line}{String of length 1. The \code{metric} to plot as the line with
the most visual salience (solid black line). \code{NULL} defaults to
"projected".}
}
\value{
An object of class "ggplot".
}
\description{
Create a trajectory plot
}
\examples{
library(dplyr, warn.conflicts = FALSE)

# `data` must meet documented "Requirements"
data <- market_share \%>\%
  filter(
    sector == "power",
    technology == "renewablescap",
    region == "global",
    scenario_source == "demo_2020"
  )
plot_trajectoryX(data)

plot_trajectoryX(data, normalize = FALSE)
}
\seealso{
\link{market_share}.

Other functions with minimal arguments: 
\code{\link{plot_techmixX}()},
\code{\link{plot_timelineX}()}
}
\concept{functions with minimal arguments}
