% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_techmix.R
\name{prep_techmix}
\alias{prep_techmix}
\alias{prep_techmixB}
\title{Prepare the output of \code{r2dii.analysis::target_market_share()} for \code{plot_timeline()}}
\usage{
prep_techmix(
  data,
  value = "technology_share",
  sector_filter = c("automotive", "aviation", "cement", "oil and gas", "shipping",
    "steel", "power"),
  years_filter = NULL,
  region_filter = "global",
  scenario_source_filter = NULL,
  scenario_filter = NULL
)

prep_techmixB(data, value = "technology_share")
}
\arguments{
\item{data}{Data frame like the output of \code{r2dii.analysis::target_market_share()}.}

\item{value}{String of length 1. The name of the column holding the value to
plot.}

\item{sector_filter}{String of length 1. Sector to pick from the \code{data}.}

\item{years_filter}{Numeric vector of length 2. Range of years to plot.}

\item{region_filter}{String of length 1. Region to pick from the \code{data}.}

\item{scenario_source_filter}{String of length 1. Value of the column
\code{scenario_source} to pick from the \code{data}.}

\item{scenario_filter}{String of length 1. Scenario to pick from the \code{data}.}
}
\description{
Prepare the output of \code{r2dii.analysis::target_market_share()} for \code{plot_timeline()}
}
\examples{
prep_techmix(
  market_share,
  sector_filter = "power",
  years_filter = c(2020, 2025),
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  scenario_filter = "sds",
  value = "technology_share"
)
library(dplyr)

data <- market_share \%>\%
  filter(
    dplyr::between(year, 2020, 2025),
    scenario_source == "demo_2020",
    sector == "power",
    region == "global",
    metric \%in\% c("projected", "corporate_economy", "target_sds")
  )

prep <- prep_techmixB(data)
}
