% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_timeline.R
\name{prep_timeline}
\alias{prep_timeline}
\title{Prepare the output of \code{r2dii.analysis::target_sda()} for \code{plot_timeline()}}
\usage{
prep_timeline(
  data,
  value = "emission_factor_value",
  metric = "emission_factor_metric",
  extrapolate = FALSE
)
}
\arguments{
\item{data}{Data frame like the output of \code{r2dii.analysis::target_sda()}.}

\item{value}{String of length 1. The name of the column holding the value to
plot.}

\item{extrapolate}{Logical of length 1. \code{TRUE} extrapolates to match the
furthest value in the data set.}

\item{Metric}{String of length 1. The name of the column holding the metrics
to plot.}
}
\value{
Data frame with columns: year, line_name, value, extrapolated.
}
\description{
Prepare the output of \code{r2dii.analysis::target_sda()} for \code{plot_timeline()}
}
\examples{
library(dplyr)

tail(prep_timeline(sda))
tail(prep_timeline(sda, extrapolate = TRUE))

data <- sda \%>\%
  rename(
    custom_value = "emission_factor_value",
    custom_metric = "emission_factor_metric"
  )
prep_timeline(data, value = "custom_value", metric = "custom_metric")
}
\seealso{
\link{sda}.
}
