% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_trajectory.R
\name{prep_trajectory}
\alias{prep_trajectory}
\alias{prep_trajectoryB}
\title{Prepares pre-processed data for plotting a trajectory chart}
\usage{
prep_trajectory(
  data,
  sector_filter,
  technology_filter,
  region_filter,
  scenario_source_filter,
  value,
  end_year_filter = 2025,
  normalize_to_start_year = TRUE
)

prep_trajectoryB(data, value = "production", normalize = TRUE)
}
\arguments{
\item{data}{Pre-processed input data.}

\item{sector_filter}{Sector for which to filter the data (character string).}

\item{technology_filter}{Technology for which to filter the data (character
string).}

\item{region_filter}{Region for which to filter the data (character string).}

\item{scenario_source_filter}{Scenario source for which to filter the data
(character string).}

\item{value, value}{The name of the value to be plotted in the trajectory chart
(character string).}

\item{end_year_filter}{Cut-off year for the chart (an integer).}

\item{normalize_to_start_year, normalize}{Logical of length-1. \code{TRUE}
normalized to start year.}
}
\value{
A data frame.
}
\description{
Prepares pre-processed data for plotting a trajectory chart
}
\examples{
prep_trajectory(
  market_share,
  sector_filter = "power",
  technology_filter = "oilcap",
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  value = "production"
)
library(dplyr)

data <- market_share \%>\%
  filter(
    technology == "oilcap",
    region == "global",
    scenario_source == "demo_2020",
    year <= 2025,
    sector == "power"
  )

prep_trajectoryB(data)
}
