% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_trajectory.R
\name{prep_trajectory}
\alias{prep_trajectory}
\title{Prepare the output of \code{r2dii.analysis::target_market_share()} for \code{plot_trajectory()}}
\usage{
prep_trajectory(
  data,
  sector_filter,
  technology_filter,
  region_filter,
  scenario_source_filter,
  value = "production",
  metric = "metric",
  end_year_filter = 2025,
  normalize = TRUE
)
}
\arguments{
\item{data}{Data frame like the output of \code{r2dii.analysis::target_market_share()}.}

\item{sector_filter}{String of length 1. Sector to pick from the \code{data}.}

\item{technology_filter}{String of length 1. Technology to pick from the
\code{data}.}

\item{region_filter}{String of length 1. Region to pick from the \code{data}.}

\item{scenario_source_filter}{String of length 1. Value of the column
\code{scenario_source} to pick from the \code{data}.}

\item{value}{String of length 1. The name of the column holding the value to
plot.}

\item{metric}{String of length 1. The name of the column holding the metrics
to plot.}

\item{end_year_filter}{Numeric of length 1. Cut-off year for the plot.}

\item{normalize}{Logical of length-1. \code{TRUE} normalizes to the start year.}
}
\value{
A data frame.
}
\description{
Prepare the output of \code{r2dii.analysis::target_market_share()} for \code{plot_trajectory()}
}
\examples{
prep_trajectory(
  market_share,
  sector_filter = "power",
  technology_filter = "oilcap",
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  value = "production"
)
}
