% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_metareport_distribution_chart}
\alias{prepare_for_metareport_distribution_chart}
\title{Prepares results data per asset type for distribution chart plot}
\usage{
prepare_for_metareport_distribution_chart(
  data_asset_type,
  sectors_filter,
  technologies_filter,
  year_filter,
  value_to_plot = "plan_carsten"
)
}
\arguments{
\item{data_asset_type}{Dataframe in the shape of
"Equity/Bonds_results_portfolio.rda" dataset from PACTA analysis output in
"30_Results" folder (dataframe).}

\item{sectors_filter}{Sector to be used for filtering (character string or a
vector of character strings).}

\item{technologies_filter}{Technologies to be used for filtering (character
string or a vector of character strings).}

\item{year_filter}{Year to be used for filtering (integer).}

\item{value_to_plot}{Variable to be used as value for plotting (character
string).}
}
\value{
Dataframe with columns investor_name, portfolio_name, value.
}
\description{
Prepares results data per asset type for distribution chart plot
}
\examples{
# TODO
}
