% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_pacta_sectors_chart}
\alias{prepare_for_pacta_sectors_chart}
\title{Prepares PACTA overview data for meta-report PACTA sectors bar chart}
\usage{
prepare_for_pacta_sectors_chart(data_overview)
}
\arguments{
\item{data_overview}{Dataframe in the shape of ".._overview_portfolio.rda"
dataset from PACTA analysis output in "30_Processed_Inputs" folder
(dataframe).}
}
\value{
Dataframe prepared for the plot.
}
\description{
Prepares PACTA overview data for meta-report PACTA sectors bar chart
}
\examples{

# TODO
}
