% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_techmix_chart.R
\name{prepare_for_techmix_chart}
\alias{prepare_for_techmix_chart}
\title{Prepares pre-processed data for plotting a tech-mix chart}
\usage{
prepare_for_techmix_chart(
  data_preprocessed,
  sector_filter,
  years_filter,
  region_filter,
  scenario_source_filter,
  scenario_filter,
  value_name
)
}
\arguments{
\item{data_preprocessed}{Pre-processed input data.}

\item{sector_filter}{Sector for which to filter the data (character string).}

\item{years_filter}{Years to plot in the graph (array of integer values).}

\item{region_filter}{Region for which to filter the data (character string).}

\item{scenario_source_filter}{Scenario source for which to filter the data
(character string).}

\item{scenario_filter}{Scenario to plot in the graph (character string).}

\item{value_name}{The name of the value to be plotted as a bar chart
(character string).}
}
\description{
Prepares pre-processed data for plotting a tech-mix chart
}
\examples{
# FIXME: Returns invisibly
out <- prepare_for_techmix_chart(
  process_input_data(example_data),
  sector_filter = "power",
  years_filter = c(2020, 2025),
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  scenario_filter = "sds",
  value_name = "technology_share"
)
out
}
