% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{prepare_for_techmix_chart}
\alias{prepare_for_techmix_chart}
\title{Prepares pre-processed data to be ready for plotting a techmix chart for a
sector}
\usage{
prepare_for_techmix_chart(
  data_preprocessed,
  sector_filter,
  years_filter,
  region_filter,
  scenario_source_filter,
  scenario_filter,
  value_name
)
}
\arguments{
\item{data_preprocessed}{pre-processed input data}

\item{sector_filter}{sector for which to filter the data (a character string)}

\item{years_filter}{years which we want to plot in the graph (an array of integer
values)}

\item{region_filter}{region for which to filter the data (a character string)}

\item{scenario_source_filter}{scenario source for which to filter the data (a
character string)}

\item{scenario_filter}{scenario to plot in the graph (a character string)}

\item{value_name}{the name of the value to be plotted as a bar chart (a
character string)}
}
\description{
Prepares pre-processed data to be ready for plotting a techmix chart for a
sector
}
