% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_trajectory_chart.R
\name{prepare_for_trajectory_chart}
\alias{prepare_for_trajectory_chart}
\title{Prepares pre-processed data for plotting a trajectory chart}
\usage{
prepare_for_trajectory_chart(
  data_preprocessed,
  sector_filter,
  technology_filter,
  region_filter,
  scenario_source_filter,
  value_name,
  end_year_filter = 2025,
  normalize_to_start_year = TRUE
)
}
\arguments{
\item{data_preprocessed}{Pre-processed input data.}

\item{sector_filter}{Sector for which to filter the data (character string).}

\item{technology_filter}{Technology for which to filter the data (character
string).}

\item{region_filter}{Region for which to filter the data (character string).}

\item{scenario_source_filter}{Scenario source for which to filter the data
(character string).}

\item{value_name}{The name of the value to be plotted in the trajectory chart
(character string).}

\item{end_year_filter}{Cut-off year for the chart (an integer).}

\item{normalize_to_start_year}{Flag indicating whether the values should be
normalized (boolean).}
}
\value{
A data frame.
}
\description{
Prepares pre-processed data for plotting a trajectory chart
}
\examples{
prepare_for_trajectory_chart(
  example_data,
  sector_filter = "power",
  technology_filter = "oilcap",
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  value_name = "production"
)
}
