% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_input_data}
\alias{process_input_data}
\title{Performs initial processing on raw input data in banks' format}
\usage{
process_input_data(data)
}
\arguments{
\item{data}{Raw input data in the banks' format.}
}
\value{
An object of class "ggolot".
}
\description{
The data is processed so that it can be used later in data filtering
functions for charts. 'metric_type' variable is added which depends on
'metric' and the 'metric' values themselves are edited for plotting purposes.
}
\examples{
# TODO
}
