% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_trajectory.R
\name{qplot_trajectory}
\alias{qplot_trajectory}
\title{Create a trajectory plot using default settings}
\usage{
qplot_trajectory(data)
}
\arguments{
\item{data}{A data frame. Requirements:
\itemize{
\item The structure must be like \link{market_share}.
\item The following columns must have a single value: \code{sector}, \code{technology},
\code{region}, \code{scenario_source}.
}}
}
\value{
An object of class "ggplot".
}
\description{
Create a trajectory plot using default settings
}
\examples{
# `data` must meet documented "Requirements"
data <- market_share \%>\%
 filter(
   sector == "power",
   technology == "renewablescap",
   region == "global",
   scenario_source == "demo_2020"

qplot_trajectory(data)
}
\seealso{
\link{market_share}.
}
