% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_trajectory_explorer.R
\name{qplot_trajectory_explorer}
\alias{qplot_trajectory_explorer}
\title{Interactive \code{qplot_trajectory} options explorer}
\usage{
qplot_trajectory_explorer(data)
}
\arguments{
\item{data}{a network description in one of numerous forms (see details)}
}
\description{
An interactive shiny widget to explore the \code{qplot_trajectory} options.
}
