% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot-package.R
\docType{package}
\name{r2dii.plot-package}
\alias{r2dii.plot}
\alias{r2dii.plot-package}
\title{r2dii.plot: Visualize Climate Scenario Alignment of Corporate Lending
    Portfolios}
\description{
These tools help you to assess if a corporate lending
    portfolio aligns with climate goals. They visualize key climate
    indicators attributed to the portfolio (e.g. production, emission
    factors), and calculate alignment targets based on climate scenarios.
    They implement in R the last step of the free software 'PACTA' (Paris
    Agreement Capital Transition Assessment;
    <https://2degrees-investing.org/>). Financial institutions use 'PACTA'
    to study how their capital allocation decisions align with climate
    change mitigation goals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.plot}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.plot/issues}
}

}
\author{
\strong{Maintainer}: Monika Furdyna \email{maurolepore@gmail.com}

}
\keyword{internal}
