% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot-package.R
\docType{package}
\name{r2dii.plot-package}
\alias{r2dii.plot}
\alias{r2dii.plot-package}
\title{r2dii.plot: Create Standard Static '2dii' Plots}
\description{
Allows the user to create standardized '2dii' graphs, like
    trajectory charts or technology mix charts, and outputs them as
    'ggplot' objects. It also contains data processing functions that
    prepare the standard PACTA data to be an input for the plotting
    functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.plot}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.plot/issues}
}

}
\author{
\strong{Maintainer}: Monika Furdyna \email{monika.furdyna@gmail.com}

}
\keyword{internal}
