% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot-package.R
\docType{package}
\name{r2dii.plot-package}
\alias{r2dii.plot}
\alias{r2dii.plot-package}
\title{r2dii.plot: Visualize the Climate Scenario Alignment of a Financial Portfolio}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create plots to visualize the alignment of a corporate lending financial portfolio to climate change scenarios based on climate indicators (production and emission intensities) across key climate relevant sectors of the 'PACTA' methodology (Paris Agreement Capital Transition Assessment; \url{https://www.transitionmonitor.com/}). Financial institutions use 'PACTA' to study how their capital allocation decisions align with climate change mitigation goals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RMI-PACTA/r2dii.plot}
  \item \url{https://rmi-pacta.github.io/r2dii.plot/}
  \item Report bugs at \url{https://github.com/RMI-PACTA/r2dii.plot/issues}
}

}
\author{
\strong{Maintainer}: Monika Furdyna \email{monika.furdyna@gmail.com} (\href{https://orcid.org/0000-0002-3728-0646}{ORCID}) [contractor]

Authors:
\itemize{
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]
  \item Alex Axthelm \email{alex@2degrees-investing.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item Rocky Mountain Institute \email{PACTA4banks@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
