% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot-package.R
\docType{package}
\name{r2dii.plot-package}
\alias{r2dii.plot}
\alias{r2dii.plot-package}
\title{r2dii.plot: Visualize the Climate Scenario Alignment of a Financial Portfolio}
\description{
Create plots to visualize the alignment of a corporate lending financial portfolio to climate change scenarios based on climate indicators (production and emission intensities) across key climate relevant sectors of the 'PACTA' methodology (Paris Agreement Capital Transition Assessment; <https://2degrees-investing.org/>). Financial institutions use 'PACTA' to study how their capital allocation decisions align with climate change mitigation goals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.plot}
  \item \url{https://2degreesinvesting.github.io/r2dii.plot/}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.plot/issues}
}

}
\author{
\strong{Maintainer}: Monika Furdyna \email{monika.furdyna@gmail.com} (\href{https://orcid.org/0000-0002-3728-0646}{ORCID}) [contractor]

Authors:
\itemize{
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]
  \item Alex Axthelm \email{alex@2degrees-investing.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
