% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2dii_colours.R
\name{r2dii_palette_colours}
\alias{r2dii_palette_colours}
\title{Outputs a data frame giving the 2DII colour palette}
\usage{
r2dii_palette_colours()
}
\description{
Outputs a data frame giving the 2DII colour palette
}
\examples{
r2dii_palette_colours()
}
