% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.R
\docType{data}
\name{sda}
\alias{sda}
\title{An example of an \code{sda}-like dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 208 rows and 4 columns.
}
\source{
\url{https://github.com/2DegreesInvesting/r2dii.plot/issues/55}.
}
\usage{
sda
}
\description{
Dataset imitating the output of \code{\link[r2dii.analysis:target_sda]{r2dii.analysis::target_sda()}}.
}
\examples{
sda
}
\seealso{
\code{\link[r2dii.analysis:target_sda]{r2dii.analysis::target_sda()}}.

Other datasets: 
\code{\link{market_share}}
}
\concept{datasets}
\keyword{datasets}
