% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_specs.R
\name{timeline_specs}
\alias{timeline_specs}
\title{Creates the default specs data frame for \code{\link[=plot_timelineA]{plot_timelineA()}}}
\usage{
timeline_specs(data)
}
\arguments{
\item{data}{Pre-processed data for the chart, with columns: year, value,
line_name.}
}
\value{
A data frame useful as the \code{specs} argument of \code{\link[=plot_timelineA]{plot_timelineA()}}.
}
\description{
Creates the default specs data frame for \code{\link[=plot_timelineA]{plot_timelineA()}}
}
\examples{
data <- prep_timelineB(sda)
timeline_specs(data)
}
\seealso{
plot_timelineA
}
