% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_title.R
\name{to_title}
\alias{to_title}
\alias{format_metric}
\alias{recode_metric_techmix}
\alias{spell_out_technology}
\title{Replicate labels produced with \verb{qplot_*()} functions}
\usage{
to_title(x)

format_metric(x)

recode_metric_techmix(x)

spell_out_technology(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\itemize{
\item \code{to_title()} converts labels like \code{\link[=qplot_emission_intensity]{qplot_emission_intensity()}}.
\item \code{format_metric()} converts labels like \code{\link[=qplot_trajectory]{qplot_trajectory()}}.
\item \code{recode_metric_techmix()} converts labels like \code{\link[=qplot_techmix]{qplot_techmix()}}.
\item \code{spell_out_technology()} converts technology labels like \code{\link[=qplot_techmix]{qplot_techmix()}}.
}
}
\examples{
to_title(c("a.string", "another_STRING"))

metric <- c("projected", "corporate_economy", "target_xyz", "else")
format_metric(metric)

recode_metric_techmix(metric)

spell_out_technology(c("gas", "ice", "coalcap", "hdv"))
}
